feat: Add alacritty aspect
Note: My config does not work with current latest release 0.6.0. I made
a pr to add optional imports. Also there was a pr that just landed in
master that expands ~ to user $HOME. This will work with the next
release 0.7.0. For now master is required. An overlay will need to be
made.