feat(git): Add git 'filter-repo' command
git 'filter-branch' was deprecated in git version 2.24. In the
deprecation message it says to use the alternative filter-repo. This is
an external [project] that needs to be installed. This is useful for
replacing git-change-author bin script with the option --mailmap

```sh
git filter-repo --mailmap mailmap
```

Note: Look at expunge alias to move functionality from filter-branch

[project]: https://github.com/newren/git-filter-repo/