fix(nvim): Set buffer and window options if exist
This change checks each option that is being processed and checks vim to
see if it is also a window or buffer option. If it is then it will also
set that option.

This fixes and issue when launching nvim the first buffer did not have
the option `number` set but all other buffers had the option set.