feat(nvim): Block packer on first install
If the packer plugin is not in
`~/.cache/nvim/site/pack/packer/opt/packer.nvim` then either it is a
fresh install or I have cleaned my cache dir. If this is the case have
neovim block on installing packages so that the rest of the config setup
does not break. This is a rough first attempt using `vim.wait`. I think
there is still some issues with the timing or the compiled script.

There is an open issue #198 on the github page
https://github.com/wbthomason/packer.nvim/ . This is something to watch
for.