feat(nvim): Add synchronous packer funcs; call on start
This commit add synchronous packer functions that can be run just like
normal packer commands. These functions can be used on bootstrapping if
packer is not installed in the `site/pack/packer/opt/packer.nvim`
folder.

These functions are also used in the justfile to make sure packer
commands are executed in order.