fix[nvim]: Remove formatoption `o` from ftplugin
For some ftplugin vim defines the `o` option for formatoption. This is
very annoying since it is overriding something that I have explicitly
disabled in my config.

I grepped though `runtime/ftplugin` searching for any languages that
they did this for and created an `after/*.vim` to make sure that it is
off.