feat(git): Add base-remote alias
This change adds a way to get the base remote for a repository. If there
is a remote named `upstream` then it will be returned else it will
default to `origin`.

This is useful when working on forks. When working on forks I am using
my fork as a place to stage branches. The source of truth is the
upstream remote and not my local origin/default.

Most aliases that use origin have been change to use base-remote where
it makes sense.