feat(nvim): Custom toggle terminal
This change was made as there is a issue with neovim and setting your
vim.o.shell to powershell. When powershell is set to your shell calling
vim.fn.termopen('powershell') will fail. This change checks to see if we
are on windows, if so then set the command to startup the terminal to be
powershell. This works around the issue.

This change also introduces the opportunity to create other togglable
terminals. Maybe a floating scratch pad?