fix(nvim): Feline responds to colorscheme changes
Feline finally responds to colorscheme changes. Everything had to be
switched over to use user defined highlight groups. The config was also
cleaned up and structured better.

Note: One of the key issues I was having was with the default highlight
group (bar color in the middle). This was fixed with a combination of
setting `default_hl` as well as having the component `default` at the
end of the left side that would set the highlight group to use
`StatusLine` as the highlight group.