fix(nyx): Build target before trying to switch
Because there might be some links to this repo created if we unlink
before trying to build the target and it fails we have lost the links
and things like neovim need to be relinked. This is a pain. A `switch`
operation should build the target, then unlink and then switch. Linking
back up if the option is passed.