feat(nvim): Use nvim-lsp-installer
Switching from using my plugin nvim-lspsync to nvim-lsp-installer as I
dont have the time or the will to work on and maintain nvim-lspsync.

This current setup lets me install my lsp servers with my normal nix way
but will be overridden if the server was installed with
nvim-lsp-installer. This also makes windows platform manageable.