feat(nvim): Add prettier, eslint & stylua to efm
I tried to use [prettier] with efm but prettier is very slow. I was
getting about a 1.5 - 2.0 second freeze when formatting on save with it.
UNACCEPTABLE! This meant that I had to package [prettierd] because it
currently does not live in nixpkgs. Figuring out how to package a yarn
project was `~~fun~~`.

To remove document formatting from `tsserver` I had to pass my
`on_attach` function to override the client capabilities. I thought that
having the capabilities in the config would work built that is not how
it is done.

This change also move stylua from the custom script I stole from `tj`,
into efm.

Also added eslint.

[prettier]: https://github.com/prettier/prettier
[prettierd]: https://github.com/fsouza/prettierd