refactor(nix): Move darwin and nixos folders into system
There are common modules and profiles between darwin and nixos systems.
This change moves both the `darwin` and `nixos` folders into a folder
called `system`. The name `system` was chosen as they both relate to
system confiugrations where as `home` does not and is system agnostic.

There is a `common` folder in system that contains common modules and
profiles. These are also sourced in the `lib.mkXXXX` functions.