feat(nix): Update default options for mkXXX funcs
This change updates the default options for the `mkHome`, `mkSystem`,
and `mkDarwin` lib functions. The functions support `config` and `user`
being either a path to the host or user file, or the name of the file.
If it is a string it will use the string as the name and look in the
correct location for the file.