fix(zsh): Add `/run/current-system/sw/bin` to PATH
There currently is an issue with zsh not adding the current-system to
teh path. I think the solution for this is to add `programs.zsh.enable`
outside of the home-manager configuration for darwin.

https://github.com/nix-community/home-manager/issues/1226

Currently there is an issue with collisions from the new nix version and
zsh when it comes to completion. This is causing me to not be able to
add `programs.zsh.enable` to my configuration.

https://github.com/LnL7/nix-darwin/issues/373
https://github.com/NixOS/nixpkgs/pull/146030

This change has not made it into unstable yet.