feat(tm): Add `jump` command to tm
This command is usefil for tmux keymappings. `tm` command manages a
`jumplist`. The keymapping would call the jump command with an
`--index`. This would check the jumplist file for that line index. If
there is a filepath present then it will use that filepath to create /
attach a tmux session with the `attach` command.

You can also edit the file with `-e` to open the file in `nvim`.