feat(nvim): Add short rev to neovim version
Because neovim is taken from a tarball it does not have the commit hash
in the version. This is hard to determine which commit neovim was built
from.

This change patches neovim to add the short rev to the version name for
when there is no git folder and just a tarball.