fix(shell): Resolves issues with using wsl and tmux
When using tmux and visual splits it would show in the far left even
though I was using the right pane.

The solution is to set `local appdata` path in
`~/.local/share/zsh/zprofile` or `~/.local/share/bash/profile`. This
change also speeds up launch time of wsl as there is no call to windows.