perf(zsh): cache zcompdump daily
Profiling zsh's startup time it was found that most of the time is spent
genreating and loading zsh's completion. The solution to this is to
cache the results of the completion dumps and only regenerate them
daily.

Zsh's `.zcompdump` file was moved to `$HOME/.cache/zsh/zcompdump` as the
default location of `$HOME/.config/zsh/` might be linked to the nix
store and be readonly. This way it is always writable.

Before:

num  calls                time                       self            name
-----------------------------------------------------------------------------------
 1)    2          32.36    16.18   53.60%     32.36    16.18   53.60%  compaudit
 2)    1          46.49    46.49   77.00%     14.13    14.13   23.40%  compinit
 3)    1           7.54     7.54   12.49%      7.44     7.44   12.32%  _zsh_highlight_load_highlighters
 4)    1           4.96     4.96    8.21%      4.96     4.96    8.21%  _zsh_highlight_bind_widgets

After:

num  calls                time                       self            name
-----------------------------------------------------------------------------------
 1)    1          13.55    13.55   49.92%     13.55    13.55   49.92%  compinit
 2)    1           7.43     7.43   27.36%      7.32     7.32   26.98%  _zsh_highlight_load_highlighters
 3)    1           4.90     4.90   18.04%      4.90     4.90   18.04%  _zsh_highlight_bind_widgets